move preprocessing out of main.go and start making test reports transient
